IDVA6-2109: Replaced old client
