--- src/test/java/uk/gov/companieshouse/acsp/manage/users/configuration/AcspApiClientConfigTest.java
@@ -1,17 +1,0 @@
-package uk.gov.companieshouse.acsp.manage.users.configuration;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-import uk.gov.companieshouse.api.handler.acspprofile.PrivateAcspProfileResourceHandler;
-
-@Tag("unit-test")
-class AcspApiClientConfigTest {
-
-    @Test
-    void getAcspResourceHAndlerIsCorrectType(){
-        Assertions.assertEquals( PrivateAcspProfileResourceHandler.class, new AcspApiClientConfig().getAcspResourceHAndler().getClass() );
-
-    }
-
-}
