+++ src/test/java/uk/gov/companieshouse/acsp/manage/users/configuration/UsersWebClientConfigTest.java
@@ -6,11 +6,11 @@ import org.junit.jupiter.api.Test;
 import org.springframework.web.reactive.function.client.WebClient;
 
 @Tag("unit-test")
-class ReactiveUsersWebClientConfigTest {
+class UsersWebClientConfigTest {
 
     @Test
     void webClientIsCreatedCorrectly(){
-        Assertions.assertTrue( WebClient.class.isAssignableFrom( new ReactiveUsersWebClientConfig().usersWebClient().getClass() ) );
+        Assertions.assertTrue( WebClient.class.isAssignableFrom( new UsersWebClientConfig().usersWebClient().getClass() ) );
     }
 
 }
