+++ src/test/java/uk/gov/companieshouse/acsp/manage/users/service/AcspProfileServiceTest.java
@@ -1,6 +1,7 @@
 package uk.gov.companieshouse.acsp.manage.users.service;
 
-import java.util.Map;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Tag;
@@ -10,55 +11,143 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.reactive.function.client.WebClientResponseException;
+import reactor.core.publisher.Mono;
 import uk.gov.companieshouse.acsp.manage.users.common.TestDataManager;
-import uk.gov.companieshouse.acsp.manage.users.rest.AcspProfileEndpoint;
-import uk.gov.companieshouse.api.acspprofile.Status;
-import uk.gov.companieshouse.api.error.ApiErrorResponseException;
-import uk.gov.companieshouse.api.handler.exception.URIValidationException;
+import uk.gov.companieshouse.acsp.manage.users.exceptions.InternalServerErrorRuntimeException;
+import uk.gov.companieshouse.acsp.manage.users.exceptions.NotFoundRuntimeException;
+import uk.gov.companieshouse.acsp.manage.users.model.AcspMembersDao;
+import uk.gov.companieshouse.api.acspprofile.AcspProfile;
 
 @ExtendWith( MockitoExtension.class )
 @Tag( "unit-test" )
 class AcspProfileServiceTest {
 
     @Mock
-    private AcspProfileEndpoint acspProfileEndpoint;
+    private WebClient acspWebClient;
 
     @InjectMocks
     private AcspProfileService acspProfileService;
 
     private static final TestDataManager testDataManager = TestDataManager.getInstance();
 
+    private void mockWebClientSuccessResponse( final String uri, final Mono<String> jsonResponse ){
+        final var requestHeadersUriSpec = Mockito.mock( WebClient.RequestHeadersUriSpec.class );
+        final var requestHeadersSpec = Mockito.mock( WebClient.RequestHeadersSpec.class );
+        final var responseSpec = Mockito.mock( WebClient.ResponseSpec.class );
 
-    @Test
-    void fetchAcspProfileRetrievesAcspProfile() throws ApiErrorResponseException, URIValidationException {
-        final var acspProfile = testDataManager.fetchAcspProfiles( "TSA001" ).getFirst();
+        Mockito.doReturn( requestHeadersUriSpec ).when( acspWebClient ).get();
+        Mockito.doReturn( requestHeadersSpec ).when( requestHeadersUriSpec ).uri( uri );
+        Mockito.doReturn( responseSpec ).when( requestHeadersSpec ).retrieve();
+        Mockito.doReturn( jsonResponse ).when( responseSpec ).bodyToMono( String.class );
+    }
 
-        Mockito.doReturn( acspProfile ).when( acspProfileEndpoint ).getAcspInfo( "TSA001" );
+    private void mockWebClientForFetchAcspProfile( final String acspNumber ) throws JsonProcessingException {
+        final var acsp = testDataManager.fetchAcspProfiles( acspNumber ).getFirst();
+        final var uri = String.format( "/authorised-corporate-service-providers/%s", acspNumber );
+        final var jsonResponse = new ObjectMapper().writeValueAsString( acsp );
+        mockWebClientSuccessResponse( uri, Mono.just( jsonResponse ) );
+    }
 
-        final var response = acspProfileService.fetchAcspProfile( "TSA001" );
-        Assertions.assertEquals( "TSA001", response.getNumber() );
-        Assertions.assertEquals( "Toy Story", response.getName() );
-        Assertions.assertEquals( Status.ACTIVE, response.getStatus() );
+    private void mockWebClientErrorResponse( final String uri, int responseCode ){
+        final var requestHeadersUriSpec = Mockito.mock( WebClient.RequestHeadersUriSpec.class );
+        final var requestHeadersSpec = Mockito.mock( WebClient.RequestHeadersSpec.class );
+        final var responseSpec = Mockito.mock( WebClient.ResponseSpec.class );
+
+        Mockito.doReturn( requestHeadersUriSpec ).when( acspWebClient ).get();
+        Mockito.doReturn( requestHeadersSpec ).when( requestHeadersUriSpec ).uri( uri );
+        Mockito.doReturn( responseSpec ).when( requestHeadersSpec ).retrieve();
+        Mockito.doReturn( Mono.error( new WebClientResponseException( responseCode, "Error", null, null, null ) ) ).when( responseSpec ).bodyToMono( String.class );
+    }
+
+    private void mockWebClientForFetchAcspProfileErrorResponse( final String acspNumber, int responseCode ){
+        final var uri = String.format( "/authorised-corporate-service-providers/%s", acspNumber );
+        mockWebClientErrorResponse( uri, responseCode );
+    }
+
+    private void mockWebClientJsonParsingError( final String uri ){
+        final var requestHeadersUriSpec = Mockito.mock( WebClient.RequestHeadersUriSpec.class );
+        final var requestHeadersSpec = Mockito.mock( WebClient.RequestHeadersSpec.class );
+        final var responseSpec = Mockito.mock( WebClient.ResponseSpec.class );
+
+        Mockito.doReturn( requestHeadersUriSpec ).when( acspWebClient ).get();
+        Mockito.doReturn( requestHeadersSpec ).when( requestHeadersUriSpec ).uri( uri );
+        Mockito.doReturn( responseSpec ).when( requestHeadersSpec ).retrieve();
+        Mockito.doReturn( Mono.just( "}{" ) ).when( responseSpec ).bodyToMono( String.class );
+    }
+
+    private void mockWebClientForFetchAcspProfileJsonParsingError( final String acspNumber ){
+        final var uri = String.format( "/authorised-corporate-service-providers/%s", acspNumber );
+        mockWebClientJsonParsingError( uri );
     }
 
     @Test
-    void fetchAcspProfilesWithNullInputThrowsNullPointerException(){
+    void fetchAcspProfileForNullOrMalformedOrNonexistentAcspReturnsNotFoundRuntimeException() {
+        mockWebClientForFetchAcspProfileErrorResponse( null, 404 );
+        Assertions.assertThrows( NotFoundRuntimeException.class, () -> acspProfileService.fetchAcspProfile( null ) );
+
+        mockWebClientForFetchAcspProfileErrorResponse( "!@£", 404 );
+        Assertions.assertThrows( NotFoundRuntimeException.class, () -> acspProfileService.fetchAcspProfile( "!@£" ) );
+
+        mockWebClientForFetchAcspProfileErrorResponse( "404Acsp", 404 );
+        Assertions.assertThrows( NotFoundRuntimeException.class, () -> acspProfileService.fetchAcspProfile( "404Acsp" ) );
+    }
+
+    @Test
+    void fetchAcspProfileWithArbitraryErrorReturnsInternalServerErrorRuntimeException() {
+        mockWebClientForFetchAcspProfileJsonParsingError( "WITA001" );
+        Assertions.assertThrows( InternalServerErrorRuntimeException.class, () -> acspProfileService.fetchAcspProfile( "WITA001" ) );
+    }
+
+    @Test
+    void fetchAcspProfileReturnsSpecifiedAcsp() throws JsonProcessingException {
+        mockWebClientForFetchAcspProfile( "WITA001" );
+        Assertions.assertEquals( "Witcher", acspProfileService.fetchAcspProfile( "WITA001" ).getName() );
+    }
+
+    @Test
+    void fetchAcspProfilesWithNullStreamThrowsNullPointerException(){
         Assertions.assertThrows( NullPointerException.class, () -> acspProfileService.fetchAcspProfiles( null ) );
     }
 
     @Test
-    void fetchAcspProfilesWithEmptyStreamReturnsEmptyMap(){
-        Assertions.assertEquals( Map.of(), acspProfileService.fetchAcspProfiles( Stream.of() ) );
+    void fetchAcspProfilesWithEmptyStreamReturnsEmptyMap() {
+        Assertions.assertEquals( 0, acspProfileService.fetchAcspProfiles( Stream.of() ).size() );
     }
 
     @Test
-    void fetchAcspProfilesRetrievesAcspProfiles() throws ApiErrorResponseException, URIValidationException {
-        final var acspMembers = testDataManager.fetchAcspMembersDaos( "TS001", "TS002" );
-        final var acspProfile = testDataManager.fetchAcspProfiles( "TSA001" ).getFirst();
+    void fetchAcspProfilesWithStreamThatHasNonExistentAcspReturnsNotFoundRuntimeException(){
+        final var membership = new AcspMembersDao();
+        membership.setAcspNumber( "404Acsp" );
+        mockWebClientForFetchAcspProfileErrorResponse( "404Acsp", 404 );
+        Assertions.assertThrows( NotFoundRuntimeException.class, () -> acspProfileService.fetchAcspProfiles( Stream.of( membership ) ) );
+    }
 
-        Mockito.doReturn( acspProfile ).when( acspProfileEndpoint ).getAcspInfo( "TSA001" );
+    @Test
+    void fetchAcspProfilesWithStreamThatHasMalformedAcspNumberReturnsInternalServerErrorRuntimeException(){
+        final var membership = new AcspMembersDao();
+        membership.setAcspNumber( "£$@123" );
+        mockWebClientForFetchAcspProfileErrorResponse( "£$@123", 400 );
+        Assertions.assertThrows( InternalServerErrorRuntimeException.class, () -> acspProfileService.fetchAcspProfiles( Stream.of( membership ) ) );
+    }
 
-        Assertions.assertEquals( Map.of( "TSA001", acspProfile ), acspProfileService.fetchAcspProfiles( acspMembers.stream() ) );
+    @Test
+    void fetchAcspProfilesWithStreamWithArbitraryErrorReturnsInternalServerErrorRuntimeException(){
+        final var membership = testDataManager.fetchAcspMembersDaos( "WIT001" ).getFirst();
+        mockWebClientForFetchAcspProfileJsonParsingError( "WITA001" );
+        Assertions.assertThrows( InternalServerErrorRuntimeException.class, () -> acspProfileService.fetchAcspProfiles( Stream.of( membership ) ) );
+    }
+
+    @Test
+    void fetchAcspProfilesWithStreamReturnsMap() throws JsonProcessingException {
+        final var membership = testDataManager.fetchAcspMembersDaos( "WIT001" ).getFirst();
+        mockWebClientForFetchAcspProfile( "WITA001" );
+        final var acsps = acspProfileService.fetchAcspProfiles( Stream.of( membership, membership ) );
+
+        Assertions.assertEquals( 1, acsps.size() );
+        Assertions.assertTrue( acsps.containsKey( "WITA001" ) );
+        Assertions.assertTrue( acsps.values().stream().map( AcspProfile::getNumber ).toList().contains( "WITA001" ) );
     }
 
 }
